LayoutTests/svg/custom/svg-parse-overflow-1-expected.txt

CONSOLE MESSAGE: line 8: Error: Invalid value for <foreignObject> attribute y="105000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
